backend/requirements.txt
Django==3.2
psycopg2-binary==2.8.6
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
